Adapt to changes to landing page
With the recent changes to the workflow, once the user is logged in,
they are redirected back to landing page and the ready method fails
as it waits for the login button it be displayed.

This has been fixed by waiting either for the username icon to be
present or the login button to be present.