Added video reporter
The reporter is intended to be used mainly in Docker image. On local
machine, it needs to have ffmpeg installed and might or might not work
well (dependeing on the system configuration).

Use: you need to enable reporter in protractor.config.ts. (set
useVideoReporter=true) and uncomment ffmpeg instalation in
Dockerfile.builder. The video will be stored in target/videos.